randomize ids (#8)
If two svgs are combined using skunk, uniqueness of ids occurring in the output is not guaranteed. This is a problem with e.g. inkscape, which generates non-unique ids like linearGradient7. Therefore combining two inkscape svgs in a single skunk svg may result in an SVG that does not meet the specification and renders incorrectly.

this solves this by randomizing IDs